+++ cache/client.go
@@ -15,9 +15,13 @@ import (
 )
 
 const (
-	opSet      = "set"
-	opGetMulti = "getmulti"
-	opDelete   = "delete"
+	opSet            = "set"
+	opGetMulti       = "getmulti"
+	opDelete         = "delete"
+	opIncrement      = "increment"
+	opTouch          = "touch"
+	opFlush          = "flushall"
+	opCompareAndSwap = "compareswap"
 
 	reasonMaxItemSize     = "max-item-size"
 	reasonAsyncBufferFull = "async-buffer-full"
@@ -69,12 +73,16 @@ func newClientMetrics(reg prometheus.Registerer) *clientMetrics {
 	cm.operations.WithLabelValues(opGetMulti)
 	cm.operations.WithLabelValues(opSet)
 	cm.operations.WithLabelValues(opDelete)
+	cm.operations.WithLabelValues(opIncrement)
+	cm.operations.WithLabelValues(opTouch)
+	cm.operations.WithLabelValues(opCompareAndSwap)
+	cm.operations.WithLabelValues(opFlush)
 
 	cm.failures = promauto.With(reg).NewCounterVec(prometheus.CounterOpts{
 		Name: "operation_failures_total",
 		Help: "Total number of operations against cache that failed.",
 	}, []string{"operation", "reason"})
-	for _, op := range []string{opGetMulti, opSet, opDelete} {
+	for _, op := range []string{opGetMulti, opSet, opDelete, opIncrement, opFlush, opTouch, opCompareAndSwap} {
 		cm.failures.WithLabelValues(op, reasonTimeout)
 		cm.failures.WithLabelValues(op, reasonMalformedKey)
 		cm.failures.WithLabelValues(op, reasonServerError)
@@ -98,6 +106,10 @@ func newClientMetrics(reg prometheus.Registerer) *clientMetrics {
 	cm.duration.WithLabelValues(opGetMulti)
 	cm.duration.WithLabelValues(opSet)
 	cm.duration.WithLabelValues(opDelete)
+	cm.duration.WithLabelValues(opIncrement)
+	cm.duration.WithLabelValues(opFlush)
+	cm.duration.WithLabelValues(opTouch)
+	cm.duration.WithLabelValues(opCompareAndSwap)
 
 	cm.dataSize = promauto.With(reg).NewHistogramVec(prometheus.HistogramOpts{
 		Name: "operation_data_size_bytes",
@@ -110,6 +122,7 @@ func newClientMetrics(reg prometheus.Registerer) *clientMetrics {
 	)
 	cm.dataSize.WithLabelValues(opGetMulti)
 	cm.dataSize.WithLabelValues(opSet)
+	cm.dataSize.WithLabelValues(opCompareAndSwap)
 
 	return cm
 }
@@ -175,6 +188,70 @@ func (c *baseClient) setAsync(key string, value []byte, ttl time.Duration, f fun
 	}
 }
 
+func (c *baseClient) touch(ctx context.Context, key string, ttl time.Duration, f func(ctx context.Context, key string, ttl time.Duration) error) error {
+	errCh := make(chan error, 1)
+
+	enqueueErr := c.asyncQueue.submit(func() {
+		start := time.Now()
+		c.metrics.operations.WithLabelValues(opTouch).Inc()
+
+		err := f(ctx, key, ttl)
+		if err != nil {
+			level.Debug(c.logger).Log(
+				"msg", "failed to touch cache item",
+				"key", key,
+				"err", err,
+			)
+			c.trackError(opTouch, err)
+		} else {
+			c.metrics.duration.WithLabelValues(opTouch).Observe(time.Since(start).Seconds())
+		}
+		errCh <- err
+	})
+
+	if errors.Is(enqueueErr, errAsyncQueueFull) {
+		c.metrics.skipped.WithLabelValues(opTouch, reasonAsyncBufferFull).Inc()
+		level.Debug(c.logger).Log("msg", "failed to touch cache item because the async buffer is full", "err", enqueueErr, "size", c.asyncBuffSize)
+		return enqueueErr
+	}
+	// Wait for the touch operation to complete.
+	return <-errCh
+}
+
+func (c *baseClient) increment(ctx context.Context, key string, delta uint64, f func(ctx context.Context, key string, delta uint64) (uint64, error)) (uint64, error) {
+	var (
+		newValue uint64
+		err      error
+	)
+	errCh := make(chan error, 1)
+
+	enqueueErr := c.asyncQueue.submit(func() {
+		start := time.Now()
+		c.metrics.operations.WithLabelValues(opIncrement).Inc()
+
+		newValue, err = f(ctx, key, delta)
+		if err != nil {
+			level.Debug(c.logger).Log(
+				"msg", "failed to increment cache item",
+				"key", key,
+				"err", err,
+			)
+			c.trackError(opIncrement, err)
+		} else {
+			c.metrics.duration.WithLabelValues(opIncrement).Observe(time.Since(start).Seconds())
+		}
+		errCh <- err
+	})
+
+	if errors.Is(enqueueErr, errAsyncQueueFull) {
+		c.metrics.skipped.WithLabelValues(opIncrement, reasonAsyncBufferFull).Inc()
+		level.Debug(c.logger).Log("msg", "failed to increment cache item because the async buffer is full", "err", enqueueErr, "size", c.asyncBuffSize)
+		return newValue, enqueueErr
+	}
+	// Wait for the increment operation to complete.
+	return newValue, <-errCh
+}
+
 // wait submits an async task and blocks until it completes. This can be used during
 // tests to ensure that async "sets" have completed before attempting to read them.
 func (c *baseClient) wait() error {
@@ -219,6 +296,66 @@ func (c *baseClient) delete(ctx context.Context, key string, f func(ctx context.
 		return enqueueErr
 	}
 	// Wait for the delete operation to complete.
+	return <-errCh
+}
+
+func (c *baseClient) compareAndSwap(ctx context.Context, key string, value []byte, ttl time.Duration, f func(ctx context.Context, key string, value []byte, ttl time.Duration) error) error {
+	errCh := make(chan error, 1)
+
+	enqueueErr := c.asyncQueue.submit(func() {
+		start := time.Now()
+		c.metrics.operations.WithLabelValues(opCompareAndSwap).Inc()
+
+		err := f(ctx, key, value, ttl)
+		if err != nil {
+			level.Debug(c.logger).Log(
+				"msg", "failed to compareAndSwap cache item",
+				"key", key,
+				"err", err,
+			)
+			c.trackError(opCompareAndSwap, err)
+		} else {
+			c.metrics.dataSize.WithLabelValues(opCompareAndSwap).Observe(float64(len(value)))
+			c.metrics.duration.WithLabelValues(opCompareAndSwap).Observe(time.Since(start).Seconds())
+		}
+		errCh <- err
+	})
+
+	if errors.Is(enqueueErr, errAsyncQueueFull) {
+		c.metrics.skipped.WithLabelValues(opCompareAndSwap, reasonAsyncBufferFull).Inc()
+		level.Debug(c.logger).Log("msg", "failed to compareAndSwap cache item because the async buffer is full", "err", enqueueErr, "size", c.asyncBuffSize)
+		return enqueueErr
+	}
+	// Wait for the compareAndSwap operation to complete.
+	return <-errCh
+}
+
+func (c *baseClient) flushAll(ctx context.Context, f func(ctx context.Context) error) error {
+	errCh := make(chan error, 1)
+
+	enqueueErr := c.asyncQueue.submit(func() {
+		start := time.Now()
+		c.metrics.operations.WithLabelValues(opFlush).Inc()
+
+		err := f(ctx)
+		if err != nil {
+			level.Debug(c.logger).Log(
+				"msg", "failed to flush all cache",
+				"err", err,
+			)
+			c.trackError(opFlush, err)
+		} else {
+			c.metrics.duration.WithLabelValues(opFlush).Observe(time.Since(start).Seconds())
+		}
+		errCh <- err
+	})
+
+	if errors.Is(enqueueErr, errAsyncQueueFull) {
+		c.metrics.skipped.WithLabelValues(opFlush, reasonAsyncBufferFull).Inc()
+		level.Debug(c.logger).Log("msg", "failed to flush all cache because the async buffer is full", "err", enqueueErr, "size", c.asyncBuffSize)
+		return enqueueErr
+	}
+	// Wait for the flush  operation to complete.
 	return <-errCh
 }